+++ autodocodec-api-usage/default.nix
@@ -13,7 +13,7 @@
 }:
 mkDerivation {
   pname = "autodocodec-api-usage";
-  version = "0.0.0.0";
+  version = "0.0.0.1";
   src = ./.;
   libraryHaskellDepends = [
     aeson autodocodec autodocodec-openapi3 autodocodec-schema