+++ autodocodec/src/Autodocodec/Aeson/Decode.hs
@@ -29,7 +29,6 @@ import Data.HashMap.Strict (HashMap)
 import qualified Data.HashMap.Strict as HashMap
 import Data.Map (Map)
 import Data.Scientific as Scientific
-import Data.Text (Text)
 import qualified Data.Text as T
 import Data.Vector (Vector)
 import qualified Data.Vector as V
@@ -66,9 +65,16 @@ parseJSONContextVia codec_ context_ =
       BoolCodec mname -> case mname of
         Nothing -> coerce (parseJSON value :: JSON.Parser Bool)
         Just name -> coerce $ withBool (T.unpack name) pure value
-      StringCodec mname -> case mname of
-        Nothing -> coerce (parseJSON value :: JSON.Parser Text)
-        Just name -> coerce $ withText (T.unpack name) pure value
+      StringCodec mname bounds ->
+        coerce $
+          ( \f -> case mname of
+              Nothing -> parseJSON value >>= f
+              Just name -> withText (T.unpack name) f value
+          )
+            ( \s -> case checkStringBounds bounds s of
+                Left err -> fail err
+                Right s' -> pure s'
+            )
       IntegerCodec mname bounds ->
         coerce $
           ( \f -> do